Fixing security issues found by CodeQL
12:22:49.389 [main] INFO  net.sf.javaanpr.test.RecognitionIT - Accuracy:	actual: 1.0	golden:0.5463917525773195

+++ src/main/java/net/sf/javaanpr/imageanalysis/Graph.java
@@ -85,7 +85,7 @@ public class Graph {
     public float getAverageValue(int a, int b) {
         float sum = 0.0f;
         for (int i = a; i < b; i++) {
-            sum += yValues.get(i).doubleValue();
+            sum += yValues.get(i);
         }
         return sum / yValues.size();
     }
